Add component and app directories to organize files
