+++ src/components/EntrancePassCard/EntrancePassCard.js
