+++ src/components/ParkDetails/ParkDetails.css
